refactor(narwhals)!: rename substrait.dataframe module to substrait.narwhals
The module is the Narwhals integration layer, not a general DataFrame; rename it
to reflect that role and free the "DataFrame" name for the native frame.

BREAKING CHANGE: import substrait.narwhals instead of substrait.dataframe. The
module was a minimal, experimental Narwhals wrapper, so impact is expected low.
